+++ Frontend/UserTech/src/app/components/list-rented-bikes/list-rented-bikes.component.html
