+++ Frontend/UserTech/src/app/components/list-rented-bikes/list-rented-bikes.component.scss
@@ -1,0 +1,1 @@
+@import 'src/app/styles/list';
