+++ Frontend/UserTech/src/app/components/list-station-bikes/list-station-bikes.component.scss
@@ -1,4 +1,4 @@
-@import "../styles/colors";
+@import "src/app/styles/colors";
 
 .bike {
   display: block;
