+++ Frontend/UserTech/src/app/components/list-stations/list-stations.component.scss
@@ -1,4 +1,4 @@
-@import './../styles/list.scss';
+@import 'src/app/styles/list';
 
 a {
   text-decoration: none;
