+++ Frontend/UserTech/src/app/components/login-view/login-view.component.scss
@@ -1,4 +1,4 @@
-@import './../styles/formStyles.scss', './../styles/general.scss';
+@import 'src/app/styles/formStyles', 'src/app/styles/general';
 
 .container{
     margin: 0;
