+++ Frontend/UserTech/src/app/components/main/main.component.scss
