+++ Frontend/UserTech/src/app/components/rent-bike/rent-bike.component.scss
@@ -1,4 +1,4 @@
-@import "../styles/colors";
+@import "src/app/styles/colors";
 
 .bike-details{
   padding: 5px 5px 0;
