--- Frontend/UserTech/src/app/list-rented-bikes/list-rented-bikes.component.scss
@@ -1,1 +1,0 @@
-@import './../styles/list.scss';
